Added an ignore channel list for commands.
Added an ignore list for channels in the CommandHandler interface
(ignoreList) and a Yolo command. Made a static final Random instance so
that new instances don't have to be created per command. Changed all
commands in the "fun" command package to use the rand instance if
needed.

Note to all: Please change your commands to use the "rand" instance of
Random if you use a Random in the command.